Use gate to access product management

+++ app/Http/Controllers/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -54,6 +55,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        Gate::authorize('view-product', $product);
+
         return view('products.show', compact('product'));
     }
 
@@ -62,6 +65,8 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        Gate::authorize('manage-product', $product);
+
         return view('products.edit', compact('product'));
     }
 
@@ -70,6 +75,8 @@ class ProductController extends Controller
      */
     public function update(Request $request,  Product $product)
     {
+        Gate::authorize('manage-product', $product);
+
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
@@ -91,6 +98,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('manage-product', $product);
+
         $product->delete();
 
         return redirect()
